v3.1.5 - a pulse that swells instead of stepping, and downloads that breathe

+++ src/Halo.App/Widgets/Fx.cs
@@ -69,7 +69,7 @@ internal static class Fx
     }
 
     public static void Glow(Graphics g, int w, int h, float fade, float cx, float cy,
-        float rx, float ry, int alpha, Color accent)
+        float rx, float ry, float alpha, Color accent)
     {
         if (accent == White || fade <= 0.01f) return;
         using var clip = PillClip(w, h);
@@ -126,7 +126,7 @@ internal static class Fx
             var oldG = g.Clip;
             g.SetClip(new RectangleF(0, 0, fill, h), CombineMode.Intersect);
             Glow(g, w, h, fade, fill * 0.45f, h * 0.44f, Math.Max(fill, h * 1.2f), h * 1.9f,
-                 (int)(16 * strength * lit), accent);
+                 16 * strength * lit, accent);
             g.Clip = oldG;
         }
 
@@ -184,7 +184,7 @@ internal static class Fx
             var oldG = g.Clip;
             g.SetClip(new RectangleF(0, 0, fill, h), CombineMode.Intersect);
             Glow(g, w, h, fade, fill, h / 2f, h * 1.1f, h * 1.45f,
-                 (int)(13 * strength * lit), accent);
+                 13 * strength * lit, accent);
             g.Clip = oldG;
         }
     }
